41-csv-data-analyzer/requirements.txt
requests
rich
click
pandas
pyyaml
streamlit
pytest
pytest-cov
flake8
black
mypy
